feat: add @constructive-io/express-context package + wire into server
Extract reusable Express middleware for Constructive tenant context into
a new shared package (@constructive-io/express-context). This enables
standalone Express services (e.g., LLM sidecar) to reuse domain
resolution, JWT auth, pgSettings, and withPgClient without depending
on the full PostGraphile server.

New package exports:
- Shared types (ApiStructure, RlsModule, AuthSettings, etc.)
- buildPgSettings() — constructs SET LOCAL key-value pairs from API + token
- withPgClient() — tenant-scoped RLS transaction helper
- requestIdMiddleware() — UUID correlation ID (X-Request-Id or generated)
- createContextMiddleware() — composes all into req.constructive

Changes to graphql/server:
- types.ts now re-exports from express-context (backwards compatible)
- middleware/types.ts re-exports ConstructiveAPIToken from express-context
- server.ts wires requestIdMiddleware + createContextMiddleware
- llm-api.ts simplified to use req.constructive instead of manual
  getPgSettings/withRlsClient/getPgPool helpers (~100 lines removed)

Refs: constructive-io/constructive-planning#917
