+++ graphql/server/package.json
@@ -42,6 +42,7 @@
   ],
   "dependencies": {
     "@agentic-kit/ollama": "^2.0.0",
+    "@constructive-io/express-context": "workspace:^",
     "@constructive-io/csrf": "workspace:^",
     "@constructive-io/graphql-env": "workspace:^",
     "@constructive-io/graphql-types": "workspace:^",
