+++ graphql/server/src/middleware/types.ts
@@ -1,25 +1,7 @@
-import type { ApiStructure } from '../types';
+// Re-export the shared ConstructiveAPIToken type for backwards compatibility.
+// The Express Request augmentation now lives in @constructive-io/express-context.
+export type { ConstructiveAPIToken } from '@constructive-io/express-context';
 
-export type ConstructiveAPIToken = {
-  id?: string;
-  user_id?: string;
-  session_id?: string;
-  access_level?: string;
-  kind?: string;
-  [key: string]: unknown;
-};
-
-declare global {
-  namespace Express {
-    interface Request {
-      api?: ApiStructure;
-      svc_key?: string;
-      clientIp?: string;
-      databaseId?: string;
-      requestId?: string;
-      token?: ConstructiveAPIToken;
-      /** Device token from constructive_device_token cookie for trusted device tracking */
-      deviceToken?: string;
-    }
-  }
-}
+// Side-effect import: pull in the Express namespace augmentation from express-context.
+// This ensures `req.api`, `req.token`, `req.constructive`, etc. are available.
+import '@constructive-io/express-context';
